Another portable_tensor_utils.cc fix for TFLM Xtensa CI build issue.
PiperOrigin-RevId: 389966051

+++ tensorflow/lite/kernels/internal/reference/portable_tensor_utils.cc
@@ -77,8 +77,8 @@ void PortableAsymmetricQuantizeFloats(const float* values, const int size,
   const double qmin_double = kMinScale;
   const double qmax_double = kMaxScale;
   const auto minmax = std::minmax_element(values, values + size);
-  const double rmin = std::min(0.0f, *minmax.first);
-  const double rmax = std::max(0.0f, *minmax.second);
+  const double rmin = static_cast<double>(std::min(0.0f, *minmax.first));
+  const double rmax = static_cast<double>(std::max(0.0f, *minmax.second));
   if (rmin == rmax) {
     memset(quantized_values, 0, size * sizeof(int8_t));
     *scaling_factor = 1;
